🐛 fix - 테스트케이스마다 QR 생성 수정
